Allow EE to opt out of member notifications
EE wants to disable certain notifications in the GroupMember and
ProjectMember models. To do so, it defines the "skip_notification"
accessor on the Member model and uses this variable in a variety of
places. This however will result in merge conflicts any time CE changes
these lines.

To resolve this, we add `Member#skip_notification?` and use this in the
places where EE expects it to be. By default this method just returns
`false`, so the CE behaviour remains the same while allowing EE to
easily redefine it.