Refactor the upgrading from source docs
Instead of requiring one separate document for every version (which is
created by release managers), we now use a single document for both
Community Edition and Enterprise Edition. This allows developers to add
guidelines right away, instead of release managers having to determine
what to add.

Version specific upgrading instructions are placed in a separate
section. This prevents instructions from lingering around for many
versions should a developer forget to remove them. This also ensures
that all instructions are kept in a single place, instead of being
spread across different documents.

A minor downside is that CE and EE guidelines now live in the same
document, which could cause merge conflicts. Since we are working
towards a single codebase, this should only be an issue until we merge
the codebases together; something we expect to do in the coming months.

This commit also removes all old upgrading instructions, but these can
still be accessed using Git branch specific URLs (included in the
updating README).